Temporary deactivation of forced disconnections on importing blocks.
Suspect race condition on importing a block which already was generated internally. The externally imported block verification fails with a "InvalidStateRoot" error, after the block of the same # was already imported internally previously.
This is a condition which can happen frequently with HoneyBadger where all Validators generate the same block at the same time and gossip them to all other Validators.
Re-evaluate this issue after integrating the hbbft Engine in the current(2.7-based) OpenEthereum branch, chances are the error does not exist on that branch(yet).